host_tools/evaluation/requirements.txt
pyelftools
capstone
angr
pathlib
bitarray
pyserial